requirements.txt
pandas
numpy
matplotlib
prophet 
scikit-learn 
umap-learn 
langgraph
langchain
redis 